feat: move set var get var to scope
