+++ packages/variable-engine/variable-core/src/scope/scope.ts
@@ -2,7 +2,7 @@ import { DisposableCollection } from '@flowgram.ai/utils';
 
 import { type VariableEngine } from '../variable-engine';
 import { createMemo } from '../utils/memo';
-import { ASTKind, MapNode } from '../ast';
+import { ASTKind, type ASTNode, type ASTNodeJSON, MapNode } from '../ast';
 import { ScopeAvailableData, ScopeEventData, ScopeOutputData } from './datas';
 
 export interface IScopeConstructor {
@@ -116,5 +116,45 @@ export class Scope<ScopeMeta extends Record<string, any> = Record<string, any>> 
 
   get disposed(): boolean {
     return this.toDispose.disposed;
+  }
+
+  /**
+   * Sets a variable in the Scope with the default key 'outputs'.
+   *
+   * @param json - The JSON value to store.
+   * @returns The updated AST node.
+   */
+  public setVar(json: ASTNodeJSON): ASTNode;
+
+  public setVar(arg1: string | ASTNodeJSON, arg2?: ASTNodeJSON): ASTNode {
+    if (typeof arg1 === 'string' && arg2 !== undefined) {
+      return this.ast.set(arg1, arg2);
+    }
+
+    if (typeof arg1 === 'object' && arg2 === undefined) {
+      return this.ast.set('outputs', arg1);
+    }
+
+    throw new Error('Invalid arguments');
+  }
+
+  /**
+   * Retrieves a variable from the Scope by key.
+   *
+   * @param key - The key of the variable to retrieve. Defaults to 'outputs'.
+   * @returns The value of the variable, or undefined if not found.
+   */
+  public getVar(key: string = 'outputs') {
+    return this.ast.get(key);
+  }
+
+  /**
+   * Clears a variable from the Scope by key.
+   *
+   * @param key - The key of the variable to clear. Defaults to 'outputs'.
+   * @returns The updated AST node.
+   */
+  public clearVar(key: string = 'outputs') {
+    return this.ast.remove(key);
   }
 }
